General docker scripts infrastructure + adding `fetch-ledger-develop:latest` docker img (#1)
This change brings new concept for creating & using docker images. It comes with general scripting infrastructure, which is reusable (can be used as is for each docker image.

It also adds setup for the fetch-ledger-develop:latest docker image which is designed to be used for devlopment & CI of the fetch-ledger code.

Please see the README.md for detailed description of this change.