Fixing ci build (propagation of env vars to sudo environment) (#12)
Our CI build started failing since the last update from centos7 to ubuntu bionic 18.04.

The failure was caused by the fact that CI image build used `gcc` instead of `clang` (well this shouldn't be the reason for failing, however in this case it was, since our code does not compile with gcc for some reason). 

The root cause was, that `CC` and `CXX` env vars, which define the compiler selected for the build, have not been propagated to `sudo` environment, what is the change when compared to our previous image. The only explanation I have for this is that new version of `sudo` might be responsible for this somehow.

The solution is to force propagation of hosting envirnment setup in ot the `sudo` environment by providing the `-E` switch to sudo commandline.